build composer version info from binary
fix of a crash in exec() for composer -n --version not being able to
identify the json_decode() global function. unfortunately it could only
be reproduced within the build but with the exactly the same command (and
environment) not in the standard shell.

removes error messages (last line of stacktrace) in the info output for
composer version.

see as well Builder::phpExec()