make docker-test-stub more idempotent
the test harness docker-test-stub part is affected by user- and group-name
as well as timestamp data (time of composer install command).

fix is to make the tar package anonymous and set the timestamp to begin of
the UNIX epoch.

additionally gzip must be commanded to do not save the original file name
and time stamp (gzip -n).