run .travis.yml as github action
start the migration to github-actions from travis-ci.

keep as many builds as possible on github.

reduce travis-ci to the effective CI/CD pipeline.

change in .travis.yml is limited to run conditionally without secrets as
these are not configured on github yet (needed for deploy to sign the
phar file). the deployment/release part should be moved to local build
first.

port of matrix via shivammathur/setup-php, more PHP versions are available.