stabilize composer which scripts
previously the composer which/which-php scripts were shell commands that
were prone to changes in composer 2 on handling shell execution on
(interactive) terminals (tty) - regression patched in composer 2.0.7.

binding the scripts to a php functions gives much more direct access to
composer which COMPOSER_BINARY value and php executable location leaving
the shell out of the equation.

the regression test from fd82423 now passes with composer 2.0.6 version
which originally failed.

ref: fd824231e845e7cb5104b6c63ad4306e5890ea43