drop macos and windows build in github ci
putting mac os and windows to a rest.

for mac os the build works but docker needs 6+ minutes to install on the
github runner. and it is slow, the build script takes x2 / x3 the time
compared to linux (w/o docker setup time). for the docker setup some
caching would be nice, the setup is merely a general issue on the github
runner for macos (requires a license for docker-enterprise github does not
have [1]).

keeping the action/workflow files thought to fire up mac os specific
testing on occasion.

windows is not supported and enough testing has been done for a start with
github ci.

additional part of the migration to github ci is to bind experimental to
allowed failures, one of the nice features of the run travis.yml action.

[1]: https://github.com/actions/virtual-environments/issues/17