fix shell tests phar creation
when a shell test is creating a pipelines phar, use build/test.phar not
build/pipelines.phar to have it specific to tests not tainting an earlier
build of the phar file.

also always run the phar build with assertions enabled, previously this
was not the case for all shell tests.

in CI (remote ci that is), do not re-build the test phar but instead
take the build/pipelines.phar from the earlier build. this prevents
triggering the composer vendor folder poisioning as the phar builder is
not properly isolated for the non-dev autoloader generation (merely for
local development).

since version 0.0.51 this has tainted the pipelines.phar file build
on travis-ci as it has been re-build from a non-reset source tree. visible
with the added "+" to the version number. the generated .asc files are
for the same phar but build non-dirty.

ref: 5ef451590bc31efeaaa63d8589b6f06dc738a251