collect artifacts of phar build
start using artifacts, each matrix step is creating artifacts of the
phar build, then collected after the ci job into one artifact and
verifying phar file check-sums across all builds.

for that set fetch depth to 0 to get actual tags as the phar file
is being versioned (in case all the same).

retention times are low as a) the first artifacts are collected anyway
and b) the collected ones are not yet further processed (e.g. deployment
is not on github ci yet).

next to the phar-file also creating the log of building it.

NOTE: pipefail is off by default on travis-ci. this is the same for the
      run .travis.yml action.

alternative could be to use phpunit 7 not with php 7.2 but 7.1 and use
phpunit 8 with php 7.2 instead of current phpunit 7. downgrading xdebug
works so far.