fix: php-binary detection
when invoking PHP and execve(2) or similar argv[0] is not an absolute path,
the PHP 5.4+ PHP_BINARY constant is an empty string which must not be
returned.

this can be provoked with:

    $ env -i php -r 'var_dump(PHP_BINARY);'
    string(0) ""

or

    $ env -u PATH php -r 'var_dump(PHP_BINARY);'
    string(0) ""

> must be resolution against PATH, but not for argv[0]?

the fallback for PHP 5.3 is then applicable for PHP 5.4+ too, just not for
the undefined const case, but for the empty (zero-length string) case.

fixes an issue building the phar-file in a meagre container environment.

NOTE: generally many things do not work well when there is no PATH and/or
      even the whole environment is cleared out. including pip.

since: 2cf36c8e84e3e9f7a9f4dd224136c464b961b8ab