build: improve shell tests portability and output
stream line the output between running on a system and running within
a container (esp. pipelines inside pipelines (pip)).

found while comparing the logs of composer ci. tested with two differs
(meld and phpstorm) and both don't get it well, so this eases review.

suppress CDPATH output:

the change-to directory information is not considered necessary in non-
interactive test-runs. this is between systems w/ and w/o CDPATH
configured.

show phar signature (CI):

in CI (e.g. inside a pipeline, travis, github-action etc.) some shell
tests do not build but copy the previous build phar file (as the assumption
is that in the first test that builds the phar always runs before).

previously the signature of the file under test has not been shown in CI
then.

change is to show the signature of the copied file, too. this streamlines
the output between a local run and a CI log and has the benefit that the
test logs show the signature of the phar under test.

grep portability:

grep(1) -cq behaves different, busybox displays the count, coreutils
didn't.

the count is not of interest but the exit status. dropping -c in favour of
-q.

improve pipelines composer-install script:

if an existing (cached) package, verify the package version and only
self-update if there is a version mismatch. previously it did always
self-update.

prevents an additional message in logs when the package already is the
version to install which can be irritating.