build: phar builder: UTC as default timezone
when building with a date.timezone on an un- or misconfigured php runtime,
an error not seen for years surfaced to not rely on the system timezone
settings as being unsafe.

as the timestamp modification is using strtotime to parse the passed string
to a UNIX epoch timestamp, this warning can be triggered.

there is nothing wrong in setting the default timezone before that, as the
warning (or fatal error with DateTime) goes away then.

it remains an issue on systems where UTC is not a valid timezone (for some
reason) and will give errors then. these should be visible as the error-
suppression operator is not in use.