build: phar builder: remove must not fail
if the remove pattern is ineffective, give notice but do not turn it into
an error. the build can succeed fine as everything has already been removed.

but still error if there is nothing to remove from.

this keeps edge-cases out, e.g. deviation in build/staging when the
development docker client stub tar package has not been created.