build: phar builder: fatal hard w/o git
got a 127 exit as git not found. gives two times the error per build.

git is used to create the build version string.

fix is to check for the error early, previously this was unnoticed and
the builder just continued.

an alternative is to provide the version as a parameter and/or a git or
git command similar git does for ssh w/ GIT_SSH and GIT_SSH_COMMAND. left
for another date, often it is easier to provide the git command within
the build container (if it is not already there).

fatal hard therefore looks reasonable on its own for now.