ci: pipelines gh-action
actionable pipeline on github actions.

integration of the php 5.3 phar build pipeline by this github action
action so that it can be used within a github action workflow.

works with phar download as it is a composite action and running
pipelines from bin/pipelines of the workspace would require to install
the project with composer first which is problematic in build situations
as it interacts with the files in the workspace which is often unwanted.