fix phar build w/ composer 2
composer version guesser deviates across build systems which poses a
problem as since composer 2 as vendor composer files contain that
different guessed version for the root package in the files:

1. vendor/composer/installed.php
2. vendor/composer/InstalledVersions.php

as the phar build includes *.php files from vendor and vendor/composer,
it inherits this deviation and taints the reproducible build.

fix is to set the COMPOSER_ROOT_VERSION as it prevents composer to detect
the version of the package.

effect is that when installed/build w/ composer 2 the installed.php and
installedVersion.php the root package data has the correct and predictable
build version:

    'root' => array (
        'pretty_version' => '0.0.56+2-g531a09+dirty',
        'version' => '0.0.56.0',
        'aliases' => array (),
        'reference' => NULL,
        'name' => 'ktomk/pipelines',
    ),

(it looses the reference thought, looks acceptable when building tags)

an alternative in 62d8fa7 was also to inject the version at build time so
that git is not required at all to build.

ref: 62d8fa7440c2b2a5912b06f306c7724e54a4ff58