mkdocs-material 7.3.6
this brings mkdocs to mkdocs 1.2.3 which is after a larger time-gap the
production test for the 1.2 series.

file:// layout broken with mkdocs for the link on the logo (docs homepage
url), thankfully mkdocs-material 7.3.6 has the

    extra:
      homepage: index.html

option. comes with the upgrade of the material theme, local workaround
was not necessary since previous mkdocs-material upgrade (7.1.1).

finally some Makefile improvements and updates.