stabilize shell tests
for out-of-tree build in a differently named directory.

additionally fix for test #2, a false positive with the assertion revealed.
grep -v (--invert-match) did also match so never failed. grep -L
(--files-without-match) did not fix it as with the more recent
distributions of grep, behaviour in exit code changed.

fix is to negate in a subshell.