fix step.clone-path support for multiple components
previously using multiple directory components in the step.clone-path
configuration parameter did fail for the symbolic link creation.

the symbolic link is use to create the clone/destination path inside the
container (via tar) with the effect having the directory the correct
user/group.

the copying of project files and directories into the container was not
affected.

this adds a routine to create a symbolic link to target directory with
parents.

finally the special case of deploying into the root folder of the
container is handled in-so-far that the clone-path directory creation
is skipped. it is not necessary and/ore requires special handling and
I don't know what happens if root is not owned by root any longer.