build: update dev dependencies
updates in the development cycle:

- friendsofphp/php-cs-fixer v3.2.1 (from v2.18.6 via v3.0.3)
- kubawerlos/php-cs-fixer-custom-fixers v3.2.0.1 (from v2.4.1 via v2.5.0)
- roave/security-advisories (dev-master -> dev-latest [1], rename only,
  updates the same)

on php 7.2 and 7.3 php-cs-fixer blocks phpunit configuration, and as
phpunit wins, php-cs-fixer has to go.

NOTE: custom fixers change is not yet in upstream main and taken from a
      fork in the meantime.

the new (temporary) composer vcs repository requires composer within any
pipeline or remote build step to have the github-oauth.github.com for the
github vcs driver. this works via composer configuration and also via the
environment, for example from the global composer config to the
environment variable:

    COMPOSER_AUTH='{"github-oauth": {"github.com": "'"$(composer config -g github-oauth.github.com)"'"}}'

or via COMPOSER_TOKEN:

    COMPOSER_TOKEN="$(composer config -g github-oauth.github.com)"
    COMPOSER_AUTH='{"github-oauth": {"github.com": "'"$COMPOSER_TOKEN"'"}}'

and for explanation:

- COMPOSER_AUTH: composer environment parameter
  https://getcomposer.org/doc/03-cli.md#composer-auth

- COMPOSER_TOKEN: environment parameter specifically for the setup-php
  GitHub action (shivammathur/setup-php@v2)
  https://github.com/shivammathur/setup-php#composer-github-oauth

[1]: https://github.com/Roave/SecurityAdvisories/commit/5ae6aa6482b72946bc0d279c3b54dc1dbeae68d8