show step annotations for manual triggers
annotate `--show` and `--show-pipelines` steps by `*M` for any step with a
manual trigger (`trigger: manual`).

streamline display of unnamed steps as `no-name` to prevent hiding steps
with `--show-pipelines`.

for that:

extract file-info modules. the file-shower should concentrate on traversing
and arranging the information in tables, not so much about data-point
string acquisition and formatting.

let file-info do that.