extract tar-copier
extract the split tar based docker cp operations to get files into a
running container (from the runner).

extracted functionality is as static methods (ease of work) and marked
with "ext" prefix.

additionally wrap both extracted methods into extDeployDirectory() as it
shares the same parameters and is/was used within StepRunner in this
fashion.  put it under test to take care of regressions.

as extracting the code here leaves more room to improve the implementation,
some checks are added.

this is to prepare better git support (currently nudging for git-worktree)
but anything may go.

it falls short for other places in use and the direction is from local
fs into container fs only.

other places are artifacts and caches which then is not provisioning (so
just a reminder).