improve yaml parse error messages
previously when parsing a pipelines YAML file that could not be parsed, for
example b/c of unsupported syntax in the YAML parser, pipelines did only
notify which file and that it was invalid for pipelines.

addition is to capture error messages so that when YAML parsing fails, this
error message is extended.

report: #14