always associate file nodes
previously some classes implementing the file-node interface were allowed
to return NULL if the node was not associated to a document.

while this can be in testing, in a scenario to use the file dom, all the
nodes are associated.

refactoring is to reduce the return type to File only, removing NULL.

this reduces code at place of use as it does not need to deal with both
types at once.

additionally step-services is demoted from a file-node,  as unused there.

additionally mocking is reduced in tests using full models.  and parts of
the test-code become more re-usable as it is now more complete.

this change prepares introducing the global options node.