support images with digest
support digest for the image properties. a digest allows to clearly
specify a container image by content (content-addressable) and is a
collision-resistant hash (pin an image).

a digest can be used instead of a tag-name with the benefit to specify a
more or less exact container image (ignoring hash collision).

there is no hash verification of the digest in pipelines itself, validation
is on level of syntax/grammar only and pipelines so far relies on the
verification by the docker implementation.

details of how a hash verification would work are outlined in the Open
Container Initiative (OCI) specs [1].

when my local docker had an issue run by digest, it exited w/ status 125
and gives reasons on standard error.

it looks like my docker expects a sha256:<hash> part encoded with
lower-case hexits (a-f not A-F).

digests work on bitbucket, documented as "Pin images by digest" [2] and
named within the JSON schema (where image names are just strings of at
least one character).

[1]: https://github.com/opencontainers/image-spec/blob/main/descriptor.md#digests
[2]: https://support.atlassian.com/bitbucket-cloud/docs/use-docker-images-as-build-environments/#Pin-images-by-digest