fix cache and data directory creation mode
when creating user directories (XDG_CACHE_HOME and XDG_DATA_HOME so far),
ensure the directories are created with access for the user only.

this is different from composer.

earlier:

    drwxrwxr-x 4 user user 4096 Jul 19  2020 /home/user/.cache/pipelines
    drwxr-xr-x 5 user user 4096 Jul 30  2020 /home/user/.cache/composer

now:

    drwx------ 3 user user 4096 Apr 25 01:22 /home/user/.local/share/pipelines

liking the idea to give it least privileges. creation of files within is
not yet covered, e.g.:

    -rwxr-xr-x 1 user user 350 Apr 25 01:32 /home/user/.local/share/pipelines/static-docker/docker-42.42.1-binsh-test-stub.14e64680ff283acb7ab6ee8fc99e21ba4da3f0c5a41a92877e362f3d0afdc546

they might need group and world executable as these are mounted into the
containers.

for caches this may be different, therefore changing file-modes after
creation of the cache tar-file to `-rw-------`.